Add `channel_axis` to `LayerNorm` ctor.
This is part 1 of fixing #186. We enable users to pass `channel_axis` which in
turn controls the shape of the scale/offset parameters. Consistent with `axis`
we support passing an int (single channel), a tuple (multiple channels) or a
slice (again multiple channels).

This enables users to get equivalent behaviour to Sonnet or Keras as they
prefer:

    keras_ln = hk.LayerNorm(.., axis=axis, channel_axis=axis)
    sonnet_ln = hk.LayerNorm(.., axis=axis, channel_axis=-1)

Currently we maintain the default of `channel_axis=-1` for any value of `axis`.
In a follow up change, I will make it an error to pass `axis!=-1` and not also
explicitly pass `channel_axis`. This will force users to pick a behaviour. As
part of this I will update internal users to maintain the current default of
`channel_axis=-1` to preserve existing behaviour.

PiperOrigin-RevId: 417571943